Add ticklist and generation and route completion handling, per gym.
Every user gets a bag into which tiddlers with route ids are made.

Something tagged "climb" is a completed climb and will have a
climbtype field.

Something tagged "tickwish" is a desired climb.

When presenting a list of routes, either the entire list of routes,
or the ticklist, two statuses (a wish or a climb) is presented along
with the route info.

If a climb is a tickwish, if it is climbed clean (redpoint, flash, onsight)
then it is removed from the ticklist.

This stuff doesn't yet use authentication, so for the time being a bag
is created for GUEST to enable testing and such.